requirements.txt
pytest~=8.3.5
allure-pytest
selene~=2.0.0rc9
allure-python-commons~=2.13.5
selenium~=4.30.0
dotenv
python-dotenv